Moved common functions used in test_hw_mii_tx_only and test_hw_mii_hot_plug to hw_helpers
