+++ tests/hw_helpers.py
@@ -11,9 +11,15 @@ from pathlib import Path
 import socket
 import time
 import json
+from decimal import Decimal
+import statistics
 from scapy.all import rdpcap, Ether, Raw, wrpcap, PcapWriter
 from mii_packet import MiiPacket
 import re
+
+# Constants used in the tests
+packet_overhead = 8 + 4 + 12 # preamble, CRC and IFG
+line_speed = 100e6
 
 
 """
@@ -469,7 +475,7 @@ def mii2pcapfile(mii_packets, pcapfile="packets.pcapng"):
         byte_list = [(nibbles[i] << 4) | nibbles[i + 1] for i in range(0, len(nibbles), 2)]
         pcap_writer.write(bytes(byte_list))
 
-    with PcapWriter(pcapfile, linktype=1) as pcap_writer: 
+    with PcapWriter(pcapfile, linktype=1) as pcap_writer:
         if isinstance(mii_packets, list):
             frames = []
             for mii_packet in mii_packets:
@@ -537,6 +543,142 @@ def calc_time_diff(start_s, start_ns, end_s, end_ns):
         diff_ns += 1000000000
 
     return (diff_s*1000000000 + diff_ns)
+
+
+def load_packet_file(filename):
+    chunk_size = 6 + 6 + 2 + 4 + 4 + 8 + 8
+    structures = []
+    with open(filename, 'rb') as f:
+        while True:
+            chunk = f.read(chunk_size)
+            if not chunk:
+                break
+
+            dst = int.from_bytes(chunk[:6], byteorder='big')
+            src = int.from_bytes(chunk[6:12], byteorder='big')
+            etype = int.from_bytes(chunk[12:14], byteorder='big')
+            seqid = int.from_bytes(chunk[14:18], byteorder='little')
+            length = int.from_bytes(chunk[18:22], byteorder='little')
+            time_s = int.from_bytes(chunk[22:30], byteorder='little')
+            time_ns = int.from_bytes(chunk[30:38], byteorder='little')
+
+            structures.append([dst, src, etype, seqid, length, time_s, time_ns])
+
+    return structures
+
+def rdpcap_to_packet_summary(packets):
+    structures = []
+    for packet in packets:
+        raw_payload = bytes(packet.payload)
+        dst = int.from_bytes(raw_payload[:6], byteorder='big')
+        src = int.from_bytes(raw_payload[6:12], byteorder='big')
+        etype = int.from_bytes(raw_payload[12:14], byteorder='big')
+        seqid = int.from_bytes(raw_payload[14:18], byteorder='little')
+        length = len(raw_payload)
+        time_s, fraction = divmod(packet.time, 1)  # provided in 'Decimal' python format
+        time_s = int(time_s)
+        time_ns = int(fraction * Decimal('1e9'))
+
+        structures.append([dst, src, etype, seqid, length, time_s, time_ns])
+    return structures
+
+def parse_packet_summary(packet_summary,
+                        expected_count_lp,
+                        expected_packet_len_lp,
+                        dut_mac_address_lp,
+                        expected_packet_len_hp = 0,
+                        dut_mac_address_hp = 0,
+                        expected_bandwidth_hp = 0,
+                        start_seq_id_lp = 0,
+                        verbose = False,
+                        check_ifg = False):
+    print("Parsing packet file")
+    # get first packet time with valid source addr
+    datum = 0
+    for packet in packet_summary:
+        if packet[1] == dut_mac_address_lp or packet[1] == dut_mac_address_hp:
+            datum = int(packet[5] * 1e9 + packet[6])
+            break
+
+    errors = ""
+    expected_seqid_lp = start_seq_id_lp
+    expected_seqid_hp = 0
+    counted_lp = 0
+    counted_hp = 0
+    last_valid_packet_time = 0
+    last_length = 0 # We need this for checking IFG
+    ifgs = []
+
+    for packet in packet_summary:
+        dst = packet[0]
+        src = packet[1]
+        seqid = packet[3]
+        length = packet[4]
+        tv_s = packet[5]
+        tv_ns = packet[6]
+        packet_time = int(tv_s * 1e9 + tv_ns) - datum
+
+        if src == dut_mac_address_lp:
+            if length != expected_packet_len_lp:
+                errors += f"Incorrect LP length at seqid: {seqid}, expected: {expected_packet_len_lp} got: {length}\n"
+            if seqid != expected_seqid_lp:
+                errors += f"Missing LP seqid: {expected_seqid_lp}, got: {seqid}\n"
+                expected_seqid_lp = seqid
+            expected_seqid_lp += 1
+            counted_lp += 1
+
+        if src == dut_mac_address_hp:
+            if length != expected_packet_len_hp:
+                errors += f"Incorrect HP length at seqid: {seqid}, expected: {expected_packet_len_hp} got: {length}\n"
+            if seqid != expected_seqid_hp:
+                errors += f"Missing HP seqid: {expected_seqid_hp}, got: {seqid}\n"
+                expected_seqid_hp = seqid
+            expected_seqid_hp += 1
+            counted_hp += 1
+
+        if src == dut_mac_address_lp or src == dut_mac_address_hp:
+            if check_ifg:
+                packet_time_diff_ns = packet_time - last_valid_packet_time
+                packet_time_ns = 1e9 / line_speed * 8 * (last_length + 4 + 8) #preamble and CRC only
+                ifg_ns = packet_time_diff_ns - packet_time_ns
+                ifgs.append(ifg_ns)
+            # ensure we only count valid packets for bandwidth calc
+            last_valid_packet_time = packet_time
+            last_length = length
+
+    if expected_bandwidth_hp:
+        total_time_ns = last_valid_packet_time # Last packet time
+        num_bits_hp = counted_hp * (expected_packet_len_hp + packet_overhead) * 8
+        bits_per_second = num_bits_hp / (total_time_ns / 1e9)
+        difference_pc = abs(expected_bandwidth_hp - bits_per_second) / abs(expected_bandwidth_hp) * 100
+        allowed_tolerance_pc = 0.1 # How close HP bandwidth should be for test pass in %
+        text = f"Calculated HP thoughput: {bits_per_second:.1f}, expected throughput: {expected_bandwidth_hp:.1f}, diff: {difference_pc:.2f}% (max: {allowed_tolerance_pc:.2f}%)"
+        if difference_pc > allowed_tolerance_pc:
+            errors += text
+        if verbose:
+            print(text)
+
+    if check_ifg:
+        ifgs = ifgs[1:-10] # The first is always wrong as is the datum and last few are HP dominared with gaps as LP tx shuts down first
+        min_ifg = min(ifgs)
+        max_ifg = max(ifgs)
+        std_dev_ifg = statistics.stdev(ifgs)
+        mean_ifg = statistics.mean(ifgs)
+        counter_dict = {}
+        for ifg in ifgs:
+            counter_dict[ifg] = counter_dict.get(ifg, 0) + 1
+        print(f"IFG stats min: {min_ifg:.2f} max: {max_ifg:.2f} mean: {mean_ifg:.2f} std_dev: {std_dev_ifg:.2f}")
+        print(f"IFG instances: {counter_dict}")
+
+    if counted_lp != expected_count_lp:
+        errors += f"Did not get: {expected_count_lp} LP packets, got: {counted_lp} (dropped: {expected_count_lp-counted_lp})"
+
+    if verbose:
+        print(f"Counted {counted_lp} LP packets and {counted_hp} HP packets over {last_valid_packet_time/1e9:.2f}s")
+
+    return errors if errors != "" else None, counted_lp, counted_hp
+
+
 
 # This is just for test
 if __name__ == "__main__":
